reorganise for separate modals
